+++ mrbgems/picoruby-mbedtls/include/cmac.h
@@ -1,0 +1,6 @@
+#ifndef MBEDTLS_CMAC_DEFINED_H_
+#define MBEDTLS_CMAC_DEFINED_H_
+
+void gem_mbedtls_cmac_init(void);
+
+#endif
